Continued improvement of autotools
Kernel module install path follows DESTDIR for packaging (but not prefix
as it should be dependent on the kernel config). INSTALL_MOD_PATH still
works if a module-specific override is needed.
make dist can create a distribution tar gz that should work and not
depend on too recent autotools.
compatibility macros for older pkg-config versions (tested up to 0.26)